# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tool Guard Core Proxy API
  version: 0.5.2
  description: >-
    Local policy-decision and escalation API exposed by tg-proxy. A denied
    response is a decision for the caller to enforce; the HTTP API does not
    execute tools.
servers:
  - url: http://127.0.0.1:9090
    description: Default local listener
...
paths:
...
  /escalations:
    get:
      tags: [Escalations]
      operationId: listEscalations
      summary: List pending and resolved escalations
      responses:
        "200":
          description: Current in-memory escalation snapshot.
          content:
            application/json:
              schema:
                type: object
                required: [escalations]
                properties:
                  escalations:
                    type: array
                    items: {$ref: "#/components/schemas/Escalation"}
        "405": {$ref: "#/components/responses/PlainMethodNotAllowed"}
  /escalations/{id}:
    parameters:
      - $ref: "#/components/parameters/EscalationID"
    get:
      tags: [Escalations]
      operationId: getEscalation
      summary: Fetch one escalation
      responses:
        "200":
          description: Escalation found.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Escalation"}
        "404": {$ref: "#/components/responses/NotFound"}
        "405": {$ref: "#/components/responses/PlainMethodNotAllowed"}
...
components:
  securitySchemes:
    approverToken:
      type: http
      scheme: bearer
      description: Token configured with tg-proxy -approver-token.
  parameters:
    EscalationID:
      in: path
      name: id
      required: true
      schema: {type: string, minLength: 1}
  responses:
...
    PlainMethodNotAllowed:
      description: Method not allowed.
      content:
        text/plain:
          schema: {type: string}
...
    NotFound:
      description: Escalation not found.
      content:
        text/plain:
          schema: {type: string}
...
  schemas:
...
    ActionEnvelope:
      type: object
      description: Versioned action context evaluated immediately before tool execution.
      properties:
        envelope_id: {type: string}
        timestamp: {type: string, format: date-time}
        agent_id: {type: string}
        session_id: {type: string}
        org_id: {type: string}
        agent_version: {type: string}
        framework: {type: string}
        turn_number: {type: integer}
        department: {type: string}
        tool_name: {type: string}
        tool_server: {type: string}
        tool_group: {type: string}
        parameters: {}
        parameters_redacted: {}
        context:
          type: object
          properties:
            verified: {type: object, additionalProperties: true}
            session_state: {type: object, additionalProperties: true}
            agent_supplied: {}
          additionalProperties: true
        integration_type: {type: string}
        proxy_version: {type: string}
        tls_verified: {type: boolean}
        integrity:
          type: object
          properties:
            hash: {type: string}
            signed_by: {type: string}
          additionalProperties: true
      additionalProperties: true
    EvaluationResultFields:
      type: object
      required: [decision, action_taken, effective_mode]
      properties:
        decision: {$ref: "#/components/schemas/Decision"}
        action_taken: {$ref: "#/components/schemas/ActionTaken"}
        decision_reason: {type: string}
        effective_mode: {$ref: "#/components/schemas/PolicyMode"}
        policies_matched: {type: integer, minimum: 0}
        rules_evaluated: {type: integer, minimum: 0}
        rules_triggered: {type: integer, minimum: 0}
        rule_results:
          type: array
          items: {$ref: "#/components/schemas/RuleResult"}
        primary_citation: {$ref: "#/components/schemas/Citation"}
        is_near_miss: {type: boolean}
        suggested_response: {type: string}
        decision_receipt:
          $ref: "#/components/schemas/DecisionReceipt"
    EvaluationResult:
      allOf:
        - $ref: "#/components/schemas/EvaluationResultFields"
      unevaluatedProperties: false
...
    RuleResult:
      type: object
      required: [rule_id, rule_name, policy_id, matched, effect, citation]
      properties:
        rule_id: {type: string}
        rule_name: {type: string}
        policy_id: {type: string}
        policy_version: {type: integer}
        matched: {type: boolean}
        effect: {type: string, enum: [allow, flag, escalate, deny]}
        severity: {type: string}
        citation: {$ref: "#/components/schemas/Citation"}
        details: {type: string}
    Citation:
      type: object
      properties:
        document_id: {type: string}
        document_title: {type: string}
        section: {type: string}
        page: {type: integer}
        line: {type: integer}
        excerpt: {type: string}
    Decision:
      type: string
      enum: [allowed, flagged, escalated, denied]
    ActionTaken:
      type: string
      enum: [allowed, flagged, escalated, denied, allowed_shadow]
    PolicyMode:
      type: string
      enum: [shadow, enforcement]
    Escalation:
      type: object
      required: [id, state, created_at, expires_at, envelope, decision]
      properties:
        id: {type: string}
        state: {type: string, enum: [pending, approved, denied, expired]}
        created_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        resolved_at: {type: string, format: date-time}
        approver: {type: string}
        approver_reason: {type: string}
        envelope: {$ref: "#/components/schemas/ActionEnvelope"}
        decision: {$ref: "#/components/schemas/EvaluationResult"}
        resolution_receipt:
          $ref: "#/components/schemas/DecisionReceipt"
    DecisionReceipt:
      type: object
      additionalProperties: false
      description: >-
        A non-resolvable correlation reference to a successfully appended
        decision trace. This is a tamper-evident hash-chain reference, not a
        signed or independently authentic receipt. It is not authorization,
        a digital signature, or a proof of non-repudiation. Durability follows
        audit-sync-mode; the receipt proves an append was accepted, not that it
        is fsync-durable beyond that mode's guarantee. Consumers must use
        action_taken for the enforced outcome, especially in shadow mode.
      required:
        - receipt_version
        - trace_id
        - trace_hash
        - hash_algorithm
        - canonical_trace_version
        - integrity_model
        - decision
        - action_taken
        - timestamp
        - receipt_uri
      properties:
        receipt_version:
          type: string
          const: "1"
        trace_id:
          type: string
          minLength: 1
          description: Persisted trace identifier copied from the audit entry.
        trace_hash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
          description: SHA-256 commitment over the canonical trace bytes.
        hash_algorithm:
          type: string
          const: sha256
        canonical_trace_version:
          type: string
          const: v2
        integrity_model:
          type: string
          const: hash-chain
        decision: {$ref: "#/components/schemas/Decision"}
        action_taken: {$ref: "#/components/schemas/ActionTaken"}
        timestamp:
          type: string
          format: date-time
          description: Timestamp persisted in the referenced trace.
        issuer:
          type: string
          description: >-
            Proxy instance identity copied from signed_by when present. Its
            presence does not make this receipt a signature.
        receipt_uri:
          type: string
          pattern: "^urn:tool-guard:trace:v2:sha256:[0-9a-f]{64}$"
          description: >-
            Opaque, hash-keyed URN. It is intentionally not a fetchable URL;
            this API publishes no receipt resolver.
...